CRM-26 making sure we don't delete the logged user

+++ app/Livewire/Admin/Users/Delete.php
@@ -39,6 +39,12 @@ class Delete extends Component
     {
         $this->validate();
 
+        if ($this->user->is(auth()->user())) {
+            $this->addError('confirmation', "You can't delete yourself brow.");
+
+            return;
+        }
+
         $this->user->delete();
 
         $this->reset('showConfirmationModal');
